added views for drain adoption
