+++ AdoptADrain/AdoptADrain/Areas/Adopt/Models/StormDrain.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace AdoptADrain.Areas.Adopt.Models
+{
+    public class StormDrain
+    {
+        public int DrainStatus { get; set; }
+    }
+}
