+++ AdoptADrain/AdoptADrain/Areas/Adopt/Views/Drain/Abandon.cshtml
@@ -1,0 +1,8 @@
+﻿
+@{
+    ViewData["Title"] = "Abandon";
+    Layout = "~/Areas/Adopt/Views/Shared/_AdoptLayout.cshtml";
+}
+
+<h1>Abandon</h1>
+
