+++ AdoptADrain/AdoptADrain/Areas/Adopt/Views/Shared/_AdoptLayout.cshtml
@@ -1,0 +1,76 @@
+﻿@*
+    For more information on enabling MVC for empty projects, visit http://go.microsoft.com/fwlink/?LinkID=397860
+*@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>@ViewData["Title"] - AdoptADrain</title>
+    <environment include="Development">
+        <link href="~/bootswatch/flatly/bootstrap.css" rel="stylesheet" />
+        <link href="~/font-awesome/css/all.css" rel="stylesheet" />
+        <link rel="stylesheet" href="~/css/site.css" />
+
+    </environment>
+    <environment exclude="Development">
+
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.1.3/css/bootstrap.min.css"
+              asp-fallback-href="~/bootswatch/flatly/bootstrap.min.css"
+              asp-fallback-test-class="sr-only" asp-fallback-test-property="position" asp-fallback-test-value="absolute"
+              crossorigin="anonymous"
+              integrity="sha256-eSi1q2PG6J7g7ib17yAaWMcrr5GrtohYChqibrV7PBE=" />
+        <link href="~/font-awesome/css/all.min.css" rel="stylesheet" />
+        <link rel="stylesheet" href="~/css/site.min.css" />
+
+    </environment>
+</head>
+<body>
+    <!-- Navigation -->
+    <nav class="navbar navbar-dark navbar-expand-lg aad-primary-color fixed-top">
+        <div class="container">
+            <ul class="mr-auto m-md-auto">
+                <li class="nav-item active"><a asp-controller="Drain" asp-action="Adopt" asp-area="Adopt">Adopt a Drain</a></li>
+                <li class="nav-item active"><a asp-controller="Drain" asp-action="Abandon" asp-area="Adopt" asp-route-name="Test">Update Drain Status</a></li>
+                <li class="nav-item active"><a asp-controller="Session" asp-action="SignOut" asp-area="">Logout</a></li>
+            </ul>
+        </div>
+    </nav>
+
+    <!-- Page Content -->
+    <div class="adoptAreaContent">
+        @RenderBody()
+    </div>
+
+    <!-- Footer -->
+    <footer class="py-3">
+        <div class="container">
+            <p class="m-0 text-center text-white">&copy; @DateTime.Now.Year  Code for Tampa Bay </p>
+        </div>
+        <!-- /.container -->
+    </footer>
+    <environment include="Development">
+        <script src="~/lib/jquery/dist/jquery.js"></script>
+        <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.js"></script>
+        <script src="~/font-awesome/js/all.js"></script>
+        <script src="~/js/site.js" asp-append-version="true"></script>
+
+    </environment>
+    <environment exclude="Development">
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js"
+                asp-fallback-src="~/lib/jquery/dist/jquery.min.js"
+                asp-fallback-test="window.jQuery"
+                crossorigin="anonymous"
+                integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.1.3/js/bootstrap.bundle.min.js"
+                asp-fallback-src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"
+                asp-fallback-test="window.jQuery && window.jQuery.fn && window.jQuery.fn.modal"
+                crossorigin="anonymous"
+                integrity="sha256-E/V4cWE4qvAeO5MOhjtGtqDzPndRO1LBk8lJ/PR7CA4="></script>
+        <script src="~/font-awesome/js/all.min.js"></script>
+        <script src="~/js/site.min.js" asp-append-version="true"></script>
+
+    </environment>
+    @RenderSection("Scripts", required: false)
+</body>
+</html>
