+++ AdoptADrain/AdoptADrain/Areas/Adopt/Views/_ViewImports.cshtml
@@ -1,0 +1,5 @@
+﻿@using AdoptADrain
+@using AdoptADrain.Models
+@using Microsoft.Extensions.Configuration
+
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
