+++ AdoptADrain/AdoptADrain/Services/IStormDrainService.cs
@@ -1,0 +1,13 @@
+﻿using AdoptADrain.DomainModels;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace AdoptADrain.Services
+{
+    public interface IStormDrainService
+    {
+        Task SaveStormDrain(StormDrain drain);
+    }
+}
