+++ AdoptADrain/AdoptADrain/Services/StormDrainService.cs
@@ -1,0 +1,13 @@
+﻿using AdoptADrain.DomainModels;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace AdoptADrain.Services
+{
+    public class StormDrainService
+    {
+
+    }       
+}
