--- AdoptADrain/AdoptADrain/StormDrain.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace AdoptADrain
-{
-    public partial class StormDrain
-    {
-        public int StormDrainId { get; set; }
-        public string Name { get; set; }
-    }
-}
