+++ AdoptADrain/AdoptADrain/appsettings.json
@@ -10,7 +10,8 @@
       "SignUpSignInPolicyId": "",
       "ResetPasswordPolicyId": "",
       "EditProfilePolicyId": "",
-      "SignInPolicyId": ""
+      "SignInPolicyId": "",
+      "RedirectUri": ""
     },
     "Logging": {
       "LogLevel": {
